Modify j2 template to use namespacing 1-level deep
The `BUILD_ENV` file (aka variables.yml\@{{build_env_name}}) was defined
in a manner that flattened the overall namespace and caused errors when
substituting.

What this meant was, variables like `SERVER_URL` that appear in multiple
sections (but with different forms, like 1. no protocol 2. with slash,
etc) were being collapsed into the _last occurrence_ parsed in the YAML
definition. This has unintended consequences. This changed uses the
intended "namespace of multiple _sections_" form.

Ultimately, this template will be removed and the BUILD_ENV will become
a _single source of truth_ (which will be used to generate any other
configuration files needed in the installation, configuration &
deployment of Atmosphere).