Generate variables.ini from generic j2 template
This corrects three significant problems:
- a j2 template in dist_files
- violation of DRY principle [1]
- violation of Open/Closed principle [2]

The use of a static, hard-coded Jinja2 template in dist_files meant
the format was needlessly repeated. So we have a variables.yml file
that define a nested-dictionary structure for the variables normally
put in <app-repo>/variables.ini. If you add a new variables, you
need to add it to the <secrets-repo>/build-env/variables.yml - then
add it to the <clank>/dist_files/<app>-variables.ini.j2.

It also meant that we could not support branched environments with
new variables available only in that environment because we would
then be *forced* to branch and synchronize releases of Clank with
releases of Atmosphere (suboptimal!).

It violated the "Open" notion of the Open/Closed principle in that
you were blocked from adding a variables and it appearing in the
application without also knowing about the template in a
non-standard directory (templates for Ansible roles go in under
`templates/` within that role; semantically - a template should not
be include in a distributation files directory). So, a variable
should be able to added to the application's `variables.ini` without
the modification of Clank (period!).

[1] https://en.wikipedia.org/wiki/Don%27t_repeat_yourself
[2] https://en.wikipedia.org/wiki/Open/closed_principle