Provide roles for handling UI assets
The handling of _legacy_ Airport UI resources is split out from the
hanlding of Troposphere UI.

These terms can be confusing in that we _refer to_ "Troposphere" as the
combination of Airport & the new user interface (aka 'Troposphere UI').

This split will allow the handling of Airport (aka cf2) to be removed
when it is ultimately retired and no longer deployed.

Playbooks that deploy the user interface may need to expose a variable
to make `install-airport-ui` optional - or author a playbook that
excludes it from being called altogether.