Provide role for nginx & uWSGI configuration
This role will link in uWSGI ini configuration file for UWSGI_APP_NAME
and then run the Makefile associated with the APPLICATION it is working
with.

In the future, the work done by the Makefile will be either brought into
this role - or done by another Ansible role. This generation of config
files done by Makefile is an infrastructure concern and should be
associated with the automation scripts for installing & configuring the
Atmosphere component (either Atmosphere API or Troposphere UI).

Side-note: it appeared the `shell:` issued previously was just a string
'make', but not executing `make`:

```
TASK: [config-nginx-uwsgi | make uwsgi and nginx stuff] ****************
changed: [localhost] => (item=make)

TASK: [config-nginx-uwsgi | debug var=output.stdout_lines] *************
ok: [localhost] => {
    "var": {
            "output.stdout_lines": "{{output.stdout_lines}}"
            }
    }
```

The evidence I offer for the observation is in the above `stdout_lines`.

By altering this to `shell: make`, I got the output I would expect from
the execution of `make` within the extras/nginx directory.

My knowledge of Ansible is such that I wouldn't put significant
confidence in the above assertion. It is noted here because I have
altered the implementation that was previously present in
<app>-setup-<app> roles.