Change pip role to include extra args & additional options
Note: This incorporates usage of `pip install` from
atmosphere-setup-atmosphere to make it functionally equivalent. However,
I do not know why `{{VENV}}/bin/wheel` is being called. I checked with
Steve Gregory and he does not know the origin of that either. I include
this just so that someone backtracking through history knows that I may
have moved this over - but I *do not* understand exactly why it is
present to begin with.