Opt for official NodeSource repo
We only depend on `npm`, and not the actual nodejs runtime. But, we
still need to ensure that we're using a recent version of npm.

It turns out that the PostCSS loader included for bundling styles via
webpack was expecting iojs 2.0 or greater - and the chrislea ppa we had
installed only 0.10.37. We needed to eventually do this, so it is time
to cross that bridge.