Add Nginx team ppa repo for stable, current version (#51)
* Remove obsolete nodejs ppa

* Opt for official NodeSource repo

We only depend on `npm`, and not the actual nodejs runtime. But, we
still need to ensure that we're using a recent version of npm.

It turns out that the PostCSS loader included for bundling styles via
webpack was expecting iojs 2.0 or greater - and the chrislea ppa we had
installed only 0.10.37. We needed to eventually do this, so it is time
to cross that bridge.

* Include stable Nginx team repository

This gives Ubuntu Trusty (14.04) access to a most recent nginx version.

The LTS has nginx 1.4.6 as the stock, available package. Including this
the stable ppa from the Nginx team will give us access to nginx 1.10.*

This means the HTTP/2 module can be enable to increase the delivery
throughput of assets (images, style, typography, javascript).

* [formatting] Remove extra space