Remove unnecessary install step
This fix makes a clank install more likely to succeed. You no longer need the
latest virtualenv.

The reason we had to update the virtualenv is to install a later version of
setuptools which ansible (silently) required. This is a bug in ansible. It is
documented in the requirements.txt.