Enable New Relic (#91)
* Defined initial role to enable New Relic

Within beta and production environments, we use the New Relic agent to
observer & monitoring the performance and duration of operations within
uWSGI processes. This can only be done in those environments.

* Defined initial role to enable New Relic

Within beta and production environments, we use the New Relic agent to
observer & monitoring the performance and duration of operations within
uWSGI processes. This can only be done in those environments.

* Extended New Relic role to generate INI

* Handled New Relic monitored app_name

* Handled New Relic monitored app_name

* Added variable example in variables.yml.dist

* Added variable requirements to role README

* Restructure role to address one component at a time

Based on feedback, this removes the repeated portions and only
addresses one component (atmosphere or troposphere) in a given role
execution.

This shifts the call into `deploy_(atmo|tropo)sphere` along with
introducing tags.

* Included `ansible_managed` directive in Jinja2 template

* Corrected typo; minus 🐮'