Added playbook to upgrade to new Postgres
This is the first _utility_ playbook defined within Clank. It does not
rely on any of the variables from the various Atmosphere build-env
files used to describe a given installation environment.

It will halt if you have not provided the two require variables:
- pg_version (the target to upgrade the server to)
- pg_version_old (the existing installation)

Optionally, you can pass a list of database to apply post-install
actions and generate "database-specific" backups:

$ clank # ansible-playbook playbooks/utils/upgrade_postgres.yml \
  --flush-cache -c local -i "localhost," \
  -e "{pg_version: '9.5',  pg_version_old: '9.3', database_names: ['atmosphere', 'troposphere']}"

Thank you Snehal Parmar for this outline provided on DBA Stack Exchange
[1] http://dba.stackexchange.com/questions/60465/upgrading-from-postgres-9-1-to-9-3-on-ubuntu-server