Fixed issue with SSH system key generation on each execution (#95)
* Resolved variable definition issue

* Ensured ssh keys will be placed correctly

What we mean by "correctly" is that the files will be put in the
expected location with a "hardcoded" key name. This will allow an
installer to state their public & private key files, but they will
then be copied & renamed to `id_rsa` & `id_rsa.pub` - which may not
be *expected*.

Right now, we would need to "variable-ize" these filenames for the
template used to generate `atmosphere/settings/local.py` [1] - and
given the overall complexity and sheer number of variables, I am
choosing not to add to that right now.

[1] https://github.com/iPlantCollaborativeOpenSource/atmosphere/blob/e0295b794b3186c9891b9684fc8220e766497d1f/atmosphere/settings/local.py.j2#L114-L117

* Ensured ssh keys will be placed correctly

What we mean by "correctly" is that the files will be put in the
expected location with a "hardcoded" key name. This will allow an
installer to state their public & private key files, but they will
then be copied & renamed to `id_rsa` & `id_rsa.pub` - which may not
be *expected*.

Right now, we would need to "variable-ize" these filenames for the
template used to generate `atmosphere/settings/local.py` [1] - and
given the overall complexity and sheer number of variables, I am
choosing not to add to that right now.

[1] https://github.com/iPlantCollaborativeOpenSource/atmosphere/blob/e0295b794b3186c9891b9684fc8220e766497d1f/atmosphere/settings/local.py.j2#L114-L117