Changed variable type of `OPENSSL_SELF_SIGNED`
We previously defined OPENSSL_SELF_SIGNED as a list containing a dict.
This made it so you could pass it to a `with_items` block, but we can
more accurately model the attributes as a dict - and then make that
dict part of a YAML list passed to `with_items`. This makes using that
variable in other tasks more nature (as you don't need to index into
it in order to get values).