Added conditional combined certification generation  (#102)
* Added conditional combined cert generation

* Handled empty SSL_CERTIFICATE/BUNDLE_CERT extra-vars

When an environment will use _self-signed_ option, then we expect the
following variables can be empty:

- SSL_CERTIFICATE
- BUNDLE_CERT

Given their precedence as "extra vars", they cannot be overridden as
facts within the task, so we opt for redefining them given usage.

(cherry picked from commit 75312d5a3f0cdd2411473d37f80465c79f450584)

* Fixed cert file format error

* Changed to only do self-signed generation if key & cert missing

If the self-signed certificate & key have already be created & are
avaiable expected locations - just path reference them

Expected locations:
-  "{{ SSL_LOCATION }}/private/{{ OPENSSL_SELF_SIGNED.name }}.key"
-  "{{ SSL_LOCATION }}/certs/{{ OPENSSL_SELF_SIGNED.name }}.crt"

* Changed variable type of `OPENSSL_SELF_SIGNED`

We previously defined OPENSSL_SELF_SIGNED as a list containing a dict.
This made it so you could pass it to a `with_items` block, but we can
more accurately model the attributes as a dict - and then make that
dict part of a YAML list passed to `with_items`. This makes using that
variable in other tasks more nature (as you don't need to index into
it in order to get values).