Allow override of DJANGO_DEBUG variable
In the current approach, a defined variable within the `group_vars` for
an application (e.g. atmosphere, troposphere) has the ability to define
an **override** of a variable, but in the [local.py] section, the
definition is just a _default_ without defining it in terms of an
override namespaced variable.  This might be total gibberish, so let's
move to some examples.

If we placed copied variables.yml.dist [0] and defined the following at
line 225 [1]:

```
ENFORCING: True
```

This would be pulled into `group_vars/atmosphere`[2] definition via the
following mechanism:

```
ATMO:
    # ....
    local.py:
        # ....
	ENFORCING: "{{ ENFORCING | default(False) }}"
```

But, this approach for defintions is not used for `DJANGO_DEBUG`, see
line 98 above `ENFORCING` [2].

This commit changes `DJANGO_DEBUG` definition to be bring in the value
of `DEBUG`.

[0] https://github.com/cyverse/clank/blob/master/dist_files/variables.yml.dist
[1] https://github.com/cyverse/clank/blob/master/dist_files/variables.yml.dist#L225
[2] https://github.com/cyverse/clank/blob/master/group_vars/atmosphere#L99