Problem: Clank installs `dev_requirements.txt` by default and it should be able to use `pip-sync` to update a virtual environment.
Solution: Changed the default requirements file variable to indicate that it
should use `requirements.txt` and also added a task that will use `pip-sync`.
I enabled this `pip-sync` task for setting up Troposphere & Atmosphere.

Note: This `pip-sync` task assumes that all the requirements for a working
deployment is in `requirements.txt` and therefore needs my recent refactoring
of the `requirements.in|txt` files in Atmosphere & Troposphere.