Fix incorrect permissions for database user, allow connections on domain sockets.
Before we didn't specify where the line was inserted. Which meant that it
would get inserted at the EOF. This means that it would fall below the line:

    local all all peer

That line is very restrictive. It states:
>On domain sockets, for all databases and all users require auth via unix login credentials

So it made our line moot. Our app continued to work because it instead used
tcp with the loopback interface.

This is significant, because engaging the network stack adds about 33%
overhead compared to domain sockets.
http://stackoverflow.com/a/12101626/1213041

Finally, in order for django with postgres to use domain sockets we need:
```
DATBASE_HOST=""
```
https://docs.djangoproject.com/en/1.10/ref/settings/#host