Revert "Pull in latest ansible galaxy tls-cert role"
This reverts commit 1fbf5724ac52b9edc5f2c06f242846b88074dc90.

After talking with @c-mart, it became apparent that this was an interesting
solution but overkill. I'm reverting it rather than nuking it, because it is
already published in several places and because it may be useful to refer to
at some point in the future where the complexity is warranted.