Create init scripts
There is now a separate service for: celeryd, celerybeat, and flower. The init
scripts have been rewritten to be less general/less configurable/and much
simpler.

The celeryd and celerybeat both run as www-data, which means that tasks run
code as the same user as the python application. This is what you'd expect.
This fixes the issues where tasks were accessing the cache as the root user,
and breaking the cache for the application.

The celeryd maging worker is an exception it still runs as root.

Flower is also run as root, this can be easily fixed. It only needs root rn so
that it can read the shared private ssl key to enable https. A workaround is
to run flower on http, and use nginx as a frontend to perform the https.