Template atmosphere-ansible
Problem: atmosphere-ansible doesn't get variables from the variables.yml file
Solution: have clank template all the files that atmosphere-ansible needs

Atmosphere-ansible has been tricky to configure, because it needs to know
about each cloud and set variables specific to that cloud. We solved this
issue by storing a bunch of files in the secrets repo under
clank_init/atmosphere-ansible/ansible. However, some of those variables files
define variables that are already defined in the variables.yml file given to
clank. This makes it really easy to misconfigure a deploy, because a variable
must be changed in more than one place.