Add support for ADDITIONAL_ALLOWED_HOSTNAMES in nginx
Note:
We are no longer returning 444 when the Host header doesn't match the
server_name.

The issue is that there is not a nice way to compare the $host variable with
each server_name. The if directive is very simple and doesn't support multiple
conditions, where you could or together tests.

This is okay because in atmosphere/tropopshere django is configured to
whitelist the hostnames in ALLOWED_HOSTS