Create backups in target folder with other backups
Problem:
Backups are created in postgresql folder not along with other backups

Solution:
Move the default directory to where other backups are made

The previous approach would create backups in postgresql and then move the
backups into another director because of postgres user permissions, if you
specifed another variable.

It's simpler to just create the backups where you intend to, this is the
proper use of sudo (despite having to tell ansible not to warn about it).